requirements.txt
backoff==2.2.1
bpy==4.4.0
dd==0.6.0
fastapi==0.115.12
geojson==3.2.0
geopy==2.4.1
graphviz==0.20.3
groq==0.24.0
ipython==8.12.3
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.59
langchain_deepseek==0.1.3
langchain_groq==0.3.2
langchain_openai==0.3.16
langgraph==0.4.3
langserve==0.3.1
mathutils==3.3.0
matplotlib==3.7.2
nltk==3.9.1
numpy==2.2.5
overpy==0.7
pandas==2.2.3
Pillow==11.2.1
psipy==0.4.0
psutil==5.9.5
pydantic==2.11.4
pydantic_settings==2.9.1
pyeda==0.29.0
pyproj==3.7.1
PyQt4==4.11.4
PySDD==1.0.5
pyswip==0.3.2
pytest==8.3.5
python-dotenv==1.1.0
python-telegram-bot==22.0
python_igraph==0.11.8
retrying==1.3.4
scikit_learn==1.6.1
scipy==1.15.3
setuptools==67.7.2
Shapely==2.1.0
smopy==0.0.8
torch==2.4.0
torchvision==0.19.0
tqdm==4.65.0
typing_extensions==4.13.2
uvicorn==0.34.2
